Some progress on tests
